Consolidate 2_2 classes
