+++ api/src/main/resources/moduleApplicationContext.xml
@@ -496,11 +496,11 @@
         </property>
     </bean>
 
-    <bean id="emrDiagnosisDAO" class="org.openmrs.module.emrapi.diagnosis.EmrDiagnosisDAOImpl2_2">
+    <bean id="emrDiagnosisDAO" class="org.openmrs.module.emrapi.diagnosis.EmrDiagnosisDAOImpl">
         <property name="sessionFactory"><ref bean="dbSessionFactory"/></property>
     </bean>
 
-    <bean id="emrVisitServiceImpl" class="${project.parent.groupId}.${project.parent.artifactId}.visit.EmrVisitServiceImpl2_2">
+    <bean id="emrVisitServiceImpl" class="${project.parent.groupId}.${project.parent.artifactId}.visit.EmrVisitServiceImpl">
         <constructor-arg ref="visitService"/>
         <constructor-arg ref="visitResponseMapper"/>
         <property name="dao">
